Fix syntax highlighting of star-projected types
Previously, when encountering a star-projected type (like `List<*>`),
kotlin-mode would highlight it only to the opening `<`. This commit adds `*` to
the characters that can appear in a type so that it correctly highlights
through the closing `>`.

Tested by creating a file `test.kt` with the following contents:
```
val x: List<*> = f()
```
Prior to this commit, this would only highlight through the opening `<`, now it
highlights through the closing `>`.